+++ src/css/overlay.css
@@ -1,0 +1,35 @@
+:root {
+  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+  line-height: 1.5;
+  font-weight: 400;
+  font-size: 22px;
+
+  color-scheme: light dark;
+  color: rgba(255, 255, 255, 0.87);
+  background-color: transparent;
+
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+
+  --tooltip-bg: #303030;
+  --tooltip-color: #fff;
+  --margin: 0.5rem;
+}
+
+body, html {
+  margin: 0;
+  padding: 0;
+  width: 100dvw;
+  height: 99dvh;
+  overflow: hidden;
+  position: relative;
+  background-color: transparent !important;
+  color: white;
+  box-sizing: content-box;
+}
+
+.main-content {
+  height: 99dvh;
+}
