+++ src/lib/YoutubeFetchMessages.ts
@@ -11,11 +11,13 @@ async function fetchYoutubeMessages(videoId: string): Promise<void> {
     const data = await chatResponse.json();
 
     data.forEach((item: ChatMessage) => {
+      const randomColor = `#${Math.floor(Math.random()*16777215).toString(16)}`;
       chatStore.addMessage({
         username: item.author,
         message: item.message,
         platform: "youtube",
         uniqueId: item.messageId,
+        usernameColor: randomColor
       });
     });
   } catch (error) {
