+++ src/routes/combined-chats-overlay/+layout.svelte
@@ -1,0 +1,11 @@
+<script>
+  import '../../css/overlay.css';
+  import 'bootstrap/dist/css/bootstrap-utilities.css';
+  import 'bootstrap/dist/css/bootstrap-grid.css';
+</script>
+
+<main>
+  <div class="main-content">
+    <slot /> <!-- This is where child pages render -->
+  </div>
+</main>
