+++ src/routes/main/twitch-chat/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { page } from '$app/stores';
-    import TwitchChat from '../../components/TwitchChat.svelte';
+    import TwitchChat from '../../../components/TwitchChat.svelte';
 
   // Access the query parameters
   $: user = $page.url.searchParams.get('user');
